requirements.txt
requests==2.24.0
beautifulsoup4==4.9.3
lxml==4.6.1
Django==3.1.2
djangorestframework==3.11.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
django-mptt==0.11.0